chore: move prediction into responsibility submodule
