+++ rex_xai/input/input_data.py
@@ -6,7 +6,7 @@ import torch as tt
 from enum import Enum
 
 from rex_xai.mutants.occlusions import spectral_occlusion, context_occlusion
-from rex_xai.prediction import Prediction
+from rex_xai.responsibility.prediction import Prediction
 from rex_xai.utils.logger import logger
 from rex_xai.utils._utils import ReXDataError
