+++ rex_xai/mutants/mutant.py
@@ -11,7 +11,7 @@ except ImportError:
 
 import matplotlib.pyplot as plt
 from rex_xai.mutants.box import Box
-from rex_xai.prediction import Prediction
+from rex_xai.responsibility.prediction import Prediction
 from rex_xai.utils.logger import logger
 from rex_xai.input.input_data import Data
 from rex_xai.utils._utils import add_boundaries, set_boolean_mask_value
