+++ rex_xai/output/visualisation.py
@@ -13,7 +13,7 @@ from scipy.ndimage import center_of_mass
 from skimage.segmentation import slic
 from torch import Tensor
 
-from rex_xai.prediction import Prediction
+from rex_xai.responsibility.prediction import Prediction
 from rex_xai.input.config import CausalArgs
 from rex_xai.responsibility.resp_maps import ResponsibilityMaps
 from rex_xai.input.input_data import Data
