+++ scripts/timm_script.py
@@ -5,7 +5,7 @@ from PIL import Image
 import torch as tt
 import torch.nn.functional as F
 from rex_xai.input.input_data import Data
-from rex_xai.prediction import from_pytorch_tensor
+from rex_xai.responsibility.prediction import from_pytorch_tensor
 
 
 model = timm.create_model("resnet152.a2_in1k", pretrained=True).to("mps")
