+++ tests/scripts/pytorch_swin_v2_t.py
@@ -6,7 +6,7 @@ import torch as tt
 import torch.nn.functional as F
 from PIL import Image  # type: ignore
 from rex_xai.input.input_data import Data
-from rex_xai.prediction import from_pytorch_tensor
+from rex_xai.responsibility.prediction import from_pytorch_tensor
 
 
 model = swin_v2_t(weights="DEFAULT")
